# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Entity.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      password:
        type: string
      roleID:
        type: string
      updatedAt:
        type: string
      userName:
        type: string
    type: object
  Request.ChangePasswordRequest:
    properties:
      currentPassword:
        minLength: 8
        type: string
      newPassword:
        minLength: 8
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
...
  Request.CreateUserRequest:
    properties:
      password:
        description: |-
          password of the user
          in: string
        minLength: 8
        type: string
      username:
        description: |-
          username of the user
          in: string
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  Request.UpdateUserRequest:
    properties:
      userName:
        type: string
    type: object
  Response.GeneralResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    required:
    - data
    - error
    - message
    type: object
...
  Response.ResponseAllUsers:
    properties:
      users:
        items:
          $ref: '#/definitions/Entity.User'
        type: array
    type: object
...
host: localhost:3000
info:
  contact:
    email: team@supnex.com
    name: Supnex
    url: https://supne.com/
  description: Golang Monolithic Boilerplate
  title: Golang Monolithic Boilerplate
  version: "1.0"
paths:
...
  /user:
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: Create user request
        in: body
        name: CreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/Request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.User'
              type: object
        "400":
          description: when user exist or password < 8 character
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Create user
      tags:
      - User
  /user-role/attach/{roleId}{userId}:
    patch:
      consumes:
      - application/json
      description: Atttach role to user with roleId and userId
      parameters:
      - description: roleId
        in: path
        name: roleId
        required: true
        type: string
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Atttach role from user
      tags:
      - User-Role
  /user-role/detach/{userId}:
    patch:
      consumes:
      - application/json
      description: Dettach role from user with roleId and userId
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Dettach role from user
      tags:
      - User-Role
  /user/:
    get:
      consumes:
      - application/json
      description: Get all users return all users with pagination
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Response.ResponseAllUsers'
              type: object
        "400":
          description: when user not exist or password is incorrect
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - User
  /user/{userId}:
    get:
      consumes:
      - application/json
      description: Get user return user with id
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.User'
              type: object
        "400":
          description: when user not exist or id is incorrect
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get user
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Change user active status with id
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.User'
              type: object
        "400":
          description: when user not exist or id is incorrect
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Change user active status
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user change user fields and return user
      parameters:
      - description: update user request
        in: body
        name: updateUser
        required: true
        schema:
          $ref: '#/definitions/Request.UpdateUserRequest'
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.User'
              type: object
        "400":
          description: when user not exist or id is incorrect
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - User
  /user/{userId}/password:
    put:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: change user password request
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/Request.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/Entity.User'
              type: object
        "400":
          description: when user not exist or id is incorrect or password in incorrect
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/Response.GeneralResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"